git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/AudioEngine@33053 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/MusicDatabase.cpp
@@ -1837,7 +1837,7 @@ bool CMusicDatabase::CleanupSongsByIds(const CStdString &strSongIds)
       m_pDS->close();
       return true;
     }
-    CStdString strSongsToDelete = "(";
+    CStdString strSongsToDelete = "";
     while (!m_pDS->eof())
     { // get the full song path
       CStdString strFileName;
@@ -1862,18 +1862,21 @@ bool CMusicDatabase::CleanupSongsByIds(const CStdString &strSongIds)
       m_pDS->next();
     }
     m_pDS->close();
-    strSongsToDelete.TrimRight(",");
-    strSongsToDelete += ")";
-    // ok, now delete these songs + all references to them from the exartistsong and exgenresong tables
-    strSQL = "delete from song where idSong in " + strSongsToDelete;
-    m_pDS->exec(strSQL.c_str());
-    strSQL = "delete from exartistsong where idSong in " + strSongsToDelete;
-    m_pDS->exec(strSQL.c_str());
-    strSQL = "delete from exgenresong where idSong in " + strSongsToDelete;
-    m_pDS->exec(strSQL.c_str());
-    strSQL = "delete from karaokedata where idSong in " + strSongsToDelete;
-    m_pDS->exec(strSQL.c_str());
-    m_pDS->close();
+
+    if ( ! strSongsToDelete.IsEmpty() )
+    {
+      strSongsToDelete = "(" + strSongsToDelete.TrimRight(",") + ")";
+      // ok, now delete these songs + all references to them from the exartistsong and exgenresong tables
+      strSQL = "delete from song where idSong in " + strSongsToDelete;
+      m_pDS->exec(strSQL.c_str());
+      strSQL = "delete from exartistsong where idSong in " + strSongsToDelete;
+      m_pDS->exec(strSQL.c_str());
+      strSQL = "delete from exgenresong where idSong in " + strSongsToDelete;
+      m_pDS->exec(strSQL.c_str());
+      strSQL = "delete from karaokedata where idSong in " + strSongsToDelete;
+      m_pDS->exec(strSQL.c_str());
+      m_pDS->close();
+    }
     return true;
   }
   catch (...)
@@ -1986,7 +1989,7 @@ bool CMusicDatabase::CleanupPaths()
       return true;
     }
     // and construct a list to delete
-    CStdString deleteSQL = "delete from path where idPath in (";
+    CStdString deleteSQL;
     while (!m_pDS->eof())
     {
       // anything that isn't a parent path of a song path is to be deleted
@@ -1997,10 +2000,13 @@ bool CMusicDatabase::CleanupPaths()
       m_pDS->next();
     }
     m_pDS->close();
-    deleteSQL.TrimRight(',');
-    deleteSQL += ")";
-    // do the deletion, and drop our temp table
-    m_pDS->exec(deleteSQL.c_str());
+
+    if ( ! deleteSQL.IsEmpty() )
+    {
+      deleteSQL = "DELETE FROM path WHERE idPath IN (" + deleteSQL.TrimRight(',') + ")";
+      // do the deletion, and drop our temp table
+      m_pDS->exec(deleteSQL.c_str());
+    }
     m_pDS->exec("drop table songpaths");
     return true;
   }
