+++ xbmc/VideoDatabase.cpp
@@ -2956,18 +2956,18 @@ void CVideoDatabase::SetVideoSettings(const CStdString& strFilenameAndPath, cons
     else
     { // add the items
       m_pDS->close();
-      strSQL=PrepareSQL("insert into settings ( idFile,Deinterlace,ViewMode,ZoomAmount,PixelRatio,"
-                       "AudioStream,SubtitleStream,SubtitleDelay,SubtitlesOn,Brightness,Contrast,Gamma,"
-                       "VolumeAmplification,AudioDelay,OutputToAllSpeakers,ResumeTime,Crop,CropLeft,CropRight,CropTop,CropBottom,Sharpness,NoiseReduction,NonLinStretch)"
-                       " values (%i,%i,%i,%f,%f,%i,%i,%f,%i,%f,%f,%f,%f,%f,%i,",
-                       idFile, setting.m_InterlaceMethod, setting.m_ViewMode, setting.m_CustomZoomAmount, setting.m_CustomPixelRatio,
-                       setting.m_AudioStream, setting.m_SubtitleStream, setting.m_SubtitleDelay, setting.m_SubtitleOn,
-                       setting.m_Brightness, setting.m_Contrast, setting.m_Gamma, setting.m_VolumeAmplification, setting.m_AudioDelay,
-                       setting.m_CustomNonLinStretch);
-      CStdString strSQL2;
-      strSQL2=PrepareSQL("%i,%i,%i,%i,%i,%i,%i,%f,%f)\n", setting.m_OutputToAllSpeakers, setting.m_ResumeTime, setting.m_Crop, setting.m_CropLeft, setting.m_CropRight,
-                    setting.m_CropTop, setting.m_CropBottom, setting.m_Sharpness, setting.m_NoiseReduction);
-      strSQL += strSQL2;
+      strSQL= "INSERT INTO settings (idFile,Deinterlace,ViewMode,ZoomAmount,PixelRatio,"
+                "AudioStream,SubtitleStream,SubtitleDelay,SubtitlesOn,Brightness,"
+                "Contrast,Gamma,VolumeAmplification,AudioDelay,OutputToAllSpeakers,"
+                "ResumeTime,Crop,CropLeft,CropRight,CropTop,CropBottom,"
+                "Sharpness,NoiseReduction,NonLinStretch) "
+              "VALUES ";
+      strSQL += PrepareSQL("(%i,%i,%i,%f,%f,%i,%i,%f,%i,%f,%f,%f,%f,%f,%i,%i,%i,%i,%i,%i,%i,%f,%f,%i)",
+                           idFile, setting.m_InterlaceMethod, setting.m_ViewMode, setting.m_CustomZoomAmount, setting.m_CustomPixelRatio,
+                           setting.m_AudioStream, setting.m_SubtitleStream, setting.m_SubtitleDelay, setting.m_SubtitleOn, setting.m_Brightness,
+                           setting.m_Contrast, setting.m_Gamma, setting.m_VolumeAmplification, setting.m_AudioDelay, setting.m_OutputToAllSpeakers,
+                           setting.m_ResumeTime, setting.m_Crop, setting.m_CropLeft, setting.m_CropRight, setting.m_CropTop, setting.m_CropBottom,
+                           setting.m_Sharpness, setting.m_NoiseReduction, setting.m_CustomNonLinStretch);
       m_pDS->exec(strSQL.c_str());
     }
   }
@@ -6183,10 +6183,10 @@ void CVideoDatabase::CleanDatabase(IVideoInfoScannerObserver* pObserver, const v
       pObserver->OnStateChanged(CLEANING_UP_DATABASE);
     }
 
-    CStdString filesToDelete = "(";
-    CStdString moviesToDelete = "(";
-    CStdString episodesToDelete = "(";
-    CStdString musicVideosToDelete = "(";
+    CStdString filesToDelete = "";
+    CStdString moviesToDelete = "";
+    CStdString episodesToDelete = "";
+    CStdString musicVideosToDelete = "";
     int total = m_pDS->num_rows();
     int current = 0;
     while (!m_pDS->eof())
@@ -6234,43 +6234,39 @@ void CVideoDatabase::CleanDatabase(IVideoInfoScannerObserver* pObserver, const v
       current++;
     }
     m_pDS->close();
-    filesToDelete.TrimRight(",");
-    filesToDelete += ")";
-    // now grab them movies
-    sql = PrepareSQL("select idMovie from movie where idFile in %s",filesToDelete.c_str());
-    m_pDS->query(sql.c_str());
-    while (!m_pDS->eof())
-    {
-      moviesToDelete += m_pDS->fv(0).get_asString() + ",";
-      m_pDS->next();
-    }
-    m_pDS->close();
-    // now grab them episodes
-    sql = PrepareSQL("select idEpisode from episode where idFile in %s",filesToDelete.c_str());
-    m_pDS->query(sql.c_str());
-    while (!m_pDS->eof())
-    {
-      episodesToDelete += m_pDS->fv(0).get_asString() + ",";
-      m_pDS->next();
-    }
-    m_pDS->close();
 
-    // now grab them musicvideos
-    sql = PrepareSQL("select idMVideo from musicvideo where idFile in %s",filesToDelete.c_str());
-    m_pDS->query(sql.c_str());
-    while (!m_pDS->eof())
+    if ( ! filesToDelete.IsEmpty() )
     {
-      musicVideosToDelete += m_pDS->fv(0).get_asString() + ",";
-      m_pDS->next();
-    }
-    m_pDS->close();
+      filesToDelete.TrimRight(",");
+      // now grab them movies
+      sql = PrepareSQL("select idMovie from movie where idFile in (%s)",filesToDelete.c_str());
+      m_pDS->query(sql.c_str());
+      while (!m_pDS->eof())
+      {
+        moviesToDelete += m_pDS->fv(0).get_asString() + ",";
+        m_pDS->next();
+      }
+      m_pDS->close();
+      // now grab them episodes
+      sql = PrepareSQL("select idEpisode from episode where idFile in (%s)",filesToDelete.c_str());
+       m_pDS->query(sql.c_str());
+      while (!m_pDS->eof())
+      {
+        episodesToDelete += m_pDS->fv(0).get_asString() + ",";
+        m_pDS->next();
+      }
+      m_pDS->close();
 
-    moviesToDelete.TrimRight(",");
-    moviesToDelete += ")";
-    episodesToDelete.TrimRight(",");
-    episodesToDelete += ")";
-    musicVideosToDelete.TrimRight(",");
-    musicVideosToDelete += ")";
+      // now grab them musicvideos
+      sql = PrepareSQL("select idMVideo from musicvideo where idFile in (%s)",filesToDelete.c_str());
+      m_pDS->query(sql.c_str());
+      while (!m_pDS->eof())
+      {
+        musicVideosToDelete += m_pDS->fv(0).get_asString() + ",";
+        m_pDS->next();
+      }
+      m_pDS->close();
+    }
 
     if (progress)
     {
@@ -6278,77 +6274,91 @@ void CVideoDatabase::CleanDatabase(IVideoInfoScannerObserver* pObserver, const v
       progress->Progress();
     }
 
-    CLog::Log(LOGDEBUG, "%s Cleaning files table", __FUNCTION__);
-    sql = "delete from files where idFile in " + filesToDelete;
-    m_pDS->exec(sql.c_str());
+    if ( ! filesToDelete.IsEmpty() )
+    {
+      filesToDelete = "(" + filesToDelete + ")";
+      CLog::Log(LOGDEBUG, "%s Cleaning files table", __FUNCTION__);
+      sql = "delete from files where idFile in " + filesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning streamdetails table", __FUNCTION__);
-    sql = "delete from streamdetails where idFile in " + filesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning streamdetails table", __FUNCTION__);
+      sql = "delete from streamdetails where idFile in " + filesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning bookmark table", __FUNCTION__);
-    sql = "delete from bookmark where idFile in " + filesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning bookmark table", __FUNCTION__);
+      sql = "delete from bookmark where idFile in " + filesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning settings table", __FUNCTION__);
-    sql = "delete from settings where idFile in " + filesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning settings table", __FUNCTION__);
+      sql = "delete from settings where idFile in " + filesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning stacktimes table", __FUNCTION__);
-    sql = "delete from stacktimes where idFile in " + filesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning stacktimes table", __FUNCTION__);
+      sql = "delete from stacktimes where idFile in " + filesToDelete;
+      m_pDS->exec(sql.c_str());
+    }
 
-    CLog::Log(LOGDEBUG, "%s Cleaning movie table", __FUNCTION__);
-    sql = "delete from movie where idMovie in " + moviesToDelete;
-    m_pDS->exec(sql.c_str());
+    if ( ! moviesToDelete.IsEmpty() )
+    {
+      moviesToDelete = "(" + moviesToDelete.TrimRight(",") + ")";
 
-    CLog::Log(LOGDEBUG, "%s Cleaning actorlinkmovie table", __FUNCTION__);
-    sql = "delete from actorlinkmovie where idMovie in " + moviesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning movie table", __FUNCTION__);
+      sql = "delete from movie where idMovie in " + moviesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning directorlinkmovie table", __FUNCTION__);
-    sql = "delete from directorlinkmovie where idMovie in " + moviesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning actorlinkmovie table", __FUNCTION__);
+      sql = "delete from actorlinkmovie where idMovie in " + moviesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning writerlinkmovie table", __FUNCTION__);
-    sql = "delete from writerlinkmovie where idMovie in " + moviesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning directorlinkmovie table", __FUNCTION__);
+      sql = "delete from directorlinkmovie where idMovie in " + moviesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning genrelinkmovie table", __FUNCTION__);
-    sql = "delete from genrelinkmovie where idMovie in " + moviesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning writerlinkmovie table", __FUNCTION__);
+      sql = "delete from writerlinkmovie where idMovie in " + moviesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning countrylinkmovie table", __FUNCTION__);
-    sql = "delete from countrylinkmovie where idMovie in " + moviesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning genrelinkmovie table", __FUNCTION__);
+      sql = "delete from genrelinkmovie where idMovie in " + moviesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning studiolinkmovie table", __FUNCTION__);
-    sql = "delete from studiolinkmovie where idMovie in " + moviesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning countrylinkmovie table", __FUNCTION__);
+      sql = "delete from countrylinkmovie where idMovie in " + moviesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning episode table", __FUNCTION__);
-    sql = "delete from episode where idEpisode in " + episodesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning studiolinkmovie table", __FUNCTION__);
+      sql = "delete from studiolinkmovie where idMovie in " + moviesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning actorlinkepisode table", __FUNCTION__);
-    sql = "delete from actorlinkepisode where idEpisode in " + episodesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning setlinkmovie table", __FUNCTION__);
+      sql = "delete from setlinkmovie where idMovie in " + moviesToDelete;
+      m_pDS->exec(sql.c_str());
+    }
 
-    CLog::Log(LOGDEBUG, "%s Cleaning directorlinkepisode table", __FUNCTION__);
-    sql = "delete from directorlinkepisode where idEpisode in " + episodesToDelete;
-    m_pDS->exec(sql.c_str());
+    if ( ! episodesToDelete.IsEmpty() )
+    {
+      episodesToDelete = "(" + episodesToDelete.TrimRight(",") + ")";
 
-    CLog::Log(LOGDEBUG, "%s Cleaning writerlinkepisode table", __FUNCTION__);
-    sql = "delete from writerlinkepisode where idEpisode in " + episodesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning episode table", __FUNCTION__);
+      sql = "delete from episode where idEpisode in " + episodesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning tvshowlinkepisode table", __FUNCTION__);
-    sql = "delete from tvshowlinkepisode where idEpisode in " + episodesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning actorlinkepisode table", __FUNCTION__);
+      sql = "delete from actorlinkepisode where idEpisode in " + episodesToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning setlinkmovie table", __FUNCTION__);
-    sql = "delete from setlinkmovie where idMovie in " + moviesToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning directorlinkepisode table", __FUNCTION__);
+      sql = "delete from directorlinkepisode where idEpisode in " + episodesToDelete;
+      m_pDS->exec(sql.c_str());
+
+      CLog::Log(LOGDEBUG, "%s Cleaning writerlinkepisode table", __FUNCTION__);
+      sql = "delete from writerlinkepisode where idEpisode in " + episodesToDelete;
+      m_pDS->exec(sql.c_str());
+
+      CLog::Log(LOGDEBUG, "%s Cleaning tvshowlinkepisode table", __FUNCTION__);
+      sql = "delete from tvshowlinkepisode where idEpisode in " + episodesToDelete;
+      m_pDS->exec(sql.c_str());
+    }
 
     CLog::Log(LOGDEBUG, "Cleaning paths that don't exist and don't have content set...");
     sql = "select * from path where strContent not like ''";
@@ -6373,11 +6383,12 @@ void CVideoDatabase::CleanDatabase(IVideoInfoScannerObserver* pObserver, const v
     CLog::Log(LOGDEBUG, "%s Cleaning tvshow table", __FUNCTION__);
     sql = "delete from tvshow where idShow not in (select idShow from tvshowlinkpath)";
     m_pDS->exec(sql.c_str());
-    sql = "delete from tvshow where idShow in (select tvshow.idShow from tvshow "
-                                               "join tvshowlinkpath on tvshow.idShow=tvshowlinkpath.idShow "
-                                               "join path on path.idPath=tvshowlinkpath.idPath "
-                                               "where tvshow.idShow not in (select idShow from tvshowlinkepisode) "
-                                               "and path.strContent == '')";
+    sql = "DELETE tvshow.* FROM tvshow "
+            "JOIN tvshowlinkpath ON tvshow.idShow=tvshowlinkpath.idShow "
+            "JOIN path ON path.idPath=tvshowlinkpath.idPath "
+          "WHERE "
+            "tvshow.idShow NOT IN ( SELECT idShow from tvshowlinkepisode ) AND "
+            "path.strContent=''";
     m_pDS->exec(sql.c_str());
 
     CLog::Log(LOGDEBUG, "%s Cleaning actorlinktvshow table", __FUNCTION__);
@@ -6402,25 +6413,30 @@ void CVideoDatabase::CleanDatabase(IVideoInfoScannerObserver* pObserver, const v
     sql = "delete from movielinktvshow where idMovie not in (select distinct idMovie from movie)";
     m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning musicvideo table", __FUNCTION__);
-    sql = "delete from musicvideo where idMVideo in " + musicVideosToDelete;
-    m_pDS->exec(sql.c_str());
+    if ( ! musicVideosToDelete.IsEmpty() )
+    {
+      musicVideosToDelete = "(" + musicVideosToDelete.TrimRight(",") + ")";
 
-    CLog::Log(LOGDEBUG, "%s Cleaning artistlinkmusicvideo table", __FUNCTION__);
-    sql = "delete from artistlinkmusicvideo where idMVideo in " + musicVideosToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning musicvideo table", __FUNCTION__);
+      sql = "delete from musicvideo where idMVideo in " + musicVideosToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning directorlinkmusicvideo table" ,__FUNCTION__);
-    sql = "delete from directorlinkmusicvideo where idMVideo in " + musicVideosToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning artistlinkmusicvideo table", __FUNCTION__);
+      sql = "delete from artistlinkmusicvideo where idMVideo in " + musicVideosToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning genrelinkmusicvideo table" ,__FUNCTION__);
-    sql = "delete from genrelinkmusicvideo where idMVideo in " + musicVideosToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning directorlinkmusicvideo table" ,__FUNCTION__);
+      sql = "delete from directorlinkmusicvideo where idMVideo in " + musicVideosToDelete;
+      m_pDS->exec(sql.c_str());
 
-    CLog::Log(LOGDEBUG, "%s Cleaning studiolinkmusicvideo table", __FUNCTION__);
-    sql = "delete from studiolinkmusicvideo where idMVideo in " + musicVideosToDelete;
-    m_pDS->exec(sql.c_str());
+      CLog::Log(LOGDEBUG, "%s Cleaning genrelinkmusicvideo table" ,__FUNCTION__);
+      sql = "delete from genrelinkmusicvideo where idMVideo in " + musicVideosToDelete;
+      m_pDS->exec(sql.c_str());
+
+      CLog::Log(LOGDEBUG, "%s Cleaning studiolinkmusicvideo table", __FUNCTION__);
+      sql = "delete from studiolinkmusicvideo where idMVideo in " + musicVideosToDelete;
+      m_pDS->exec(sql.c_str());
+    }
 
     CLog::Log(LOGDEBUG, "%s Cleaning path table", __FUNCTION__);
     sql = "delete from path where idPath not in (select distinct idPath from files) and idPath not in (select distinct idPath from tvshowlinkpath) and strContent=''";
